Add state tracker
we need a bit of state tracking to know which shaders are bound at all
times.  this will be used to associate assembly output with the draw
calls.

State dumping code elsewhere in apitrace just queries this, which
could work as well, if all the queries are supported.  However, we
need to use tracking method for assembly output anyways.

Also fixed some bugs in call offset handling in render bookmarks.