Clean up static maps after retrace.
retrace has a cleanup() routine which can be called to free resources
after a retrace.  However, it fails to free resources which are
allocated in several static global data structures.

Due to this shortcoming, a trace file cannot be traced twice in a single
process.  This directly affects gtests which need to execute a minimal
trace file and check the results.