Experiments: Add Disable Draw experiment
Add a tab to hold frame experiments.  The first experiment is the
simplest: Disable the selected draw calls.

When an experiment is enabled, all UI for the frame needs to be
refreshed to take into account the modified frame:

 - Bar graph
 - Metrics table
 - Render target
 - Shaders (for Disable Draw, they will be unchanged)
 - GPU Batch

Subsequent retrace requests for these items are out of date at the point
an experiment is enabled, in the same way that Selection outdates
several the API log and other visualizations.

ExperimentId is distributed throughout the models in the same way as
selection, and they respond to that signal to request new data.  The
Retrace interface drops out of date requests based on both selection and
experiment status.