Metrics: Drop out-of date refresh requests
Clicking multiple times on the refresh button would enqueue several long
queries, if the frame is large.  This commit uses the experiment count
to drop obsolete requests.  It imposes an additional cost to refresh of
getting new render targets and shaders, but this is much better than the
very long latencies that were possible before.