State: Refactor state display
Initial implementation used ad-hoc hints to correctly display various
types of data.  With this commit, a generic implementation
incorporates the "indices" into the data model, so all state can be
displayed as a vector of items.  The state table is not re-created on
each selection, which maintains the state of the triangle widget for
collapsing directories.