Retrace: Provide an asynchronous cancellation mechanism
The initial client/server api implemented synchronous calls, with
asynchronous callbacks.  Under this scheme, it is easy to map server
responses to the client callback pointer provided to the stub.

However, the synchronous server implementation does not allow the
client to cancel requests which may have a long latency.  Retracing
textures associated with a large selection can take several minutes
due to the size of the textures.

This commit implements a new cancellation mechanism.  The server
provides a second socket thread to listen for updates to the
experiment count or the selection count.  The client sends those
counts whenever the user invalidates a prior request.  Long-latency
server commands will check the counts as they iterate through the
workload, returning prematurely if the workload result is no longer
needed by the gui.